Apply Learning of Abstract Factory Design Pattern Where added another layer in factory design pattern and make it more secure and loosely coupled. In this, We create one more abstract class named EmployeeAbstractFactroy class  having method createEmployee() of return type Employee and pass that class into method named getEmployee() having return type Employee defined in EmployeeFactory class. Now We create two implementation class of EmployeeAbstractFactory, i.e. AndroidDevFactory class and WebDevFactory class, and override the createEmployee() method and return the corresponding objects based on their implementation class.
now in main method or client class, to create object of android
developer and web developer, we call getEmployee() method of
EmployeeFactory and pass the implementation class
EmployeeAbstractFactory class and on basis of implementation class, It
will return the object of the type of employee.

Hence We implement Abstract Factory Design pattern where both factory
does not know exactly which employee type object they are creating.